+++ code-t2/src/HomeAutomation/LightScheduler.c
@@ -45,7 +45,7 @@ typedef struct
 {
     int id;
     Day day;
-    long int minuteOfDay;
+    int minuteOfDay;
     int event;
     int randomize;
     int randomMinutes;
@@ -73,7 +73,7 @@ static void resetRandomize(ScheduledLightEvent * e);
 
 #if 1 
 static void setEventSchedule(ScheduledLightEvent * event,
-        int id, Day day, long int minute, int control, int randomize)
+        int id, Day day, int minute, int control, int randomize)
 {
         event->id = id;
         event->day = day;
@@ -122,7 +122,7 @@ static ScheduledLightEvent * findUnusedEvent(void)
 #endif 
 
 #if 1 
-static void scheduleEvent(int id, Day day, long int minute, int control,
+static void scheduleEvent(int id, Day day, int minute, int control,
         int randomize)
 {
 
@@ -134,7 +134,7 @@ static void scheduleEvent(int id, Day day, long int minute, int control,
 #endif 
 
 #if 0 
-static void scheduleEvent(int id, Day day, long int minuteOfDay, int control,
+static void scheduleEvent(int id, Day day, int minuteOfDay, int control,
         int randomize)
 {
 
@@ -154,22 +154,22 @@ static void scheduleEvent(int id, Day day, long int minuteOfDay, int control,
 #endif 
 
 
-void LightScheduler_ScheduleTurnOn(int id, Day day, long int minute)
+void LightScheduler_ScheduleTurnOn(int id, Day day, int minute)
 {
     scheduleEvent(id, day, minute, TURN_ON, RANDOM_OFF);
 }
 
-void LightScheduler_ScheduleTurnOff(int id, int day, long int minute)
+void LightScheduler_ScheduleTurnOff(int id, int day, int minute)
 {
     scheduleEvent(id, day, minute, TURN_OFF, RANDOM_OFF);
 }
 
-static BOOL matchEvent(ScheduledLightEvent * e, int id, Day day, long int minute)
+static BOOL matchEvent(ScheduledLightEvent * e, int id, Day day, int minute)
 {
     return e->id == id && e->day == day && e->minuteOfDay == minute;
 }
 
-void LightScheduler_Randomize(int id, Day day, long int minute)
+void LightScheduler_Randomize(int id, Day day, int minute)
 {
     int i;
     ScheduledLightEvent * event = eventList;
@@ -184,7 +184,7 @@ void LightScheduler_Randomize(int id, Day day, long int minute)
     }
 }
 
-void LightScheduler_ScheduleRemove(int id, Day day, long int minute)
+void LightScheduler_ScheduleRemove(int id, Day day, int minute)
 {
     int i;
     ScheduledLightEvent * event = eventList;
@@ -213,7 +213,7 @@ static void resetRandomize(ScheduledLightEvent * event)
 }
 
 #if 0 
-static void scheduleEvent(int id, Day day, long int minuteOfDay, int control,
+static void scheduleEvent(int id, Day day, int minuteOfDay, int control,
         int randomize)
 {
     int i;
@@ -241,7 +241,7 @@ static void scheduleEvent(int id, Day day, long int minuteOfDay, int control,
 #endif 
 
 #if 0 
-static void scheduleEvent(int id, Day day, long int minuteOfDay, int control,
+static void scheduleEvent(int id, Day day, int minuteOfDay, int control,
         int randomize)
 {
     int i;
@@ -264,7 +264,7 @@ static void scheduleEvent(int id, Day day, long int minuteOfDay, int control,
 #endif 
 
 #if 0 
-static void scheduleEvent(int id, Day day, long int minuteOfDay, int control,
+static void scheduleEvent(int id, Day day, int minuteOfDay, int control,
         int randomize)
 {
     int i;
